refactor(companyLocation): singular file and constant name
