+++ studio/schemas/deskStructure.ts
@@ -17,7 +17,7 @@ import { soMeLinksID } from "./documents/socialMediaProfiles";
 import { companyInfoID } from "./documents/companyInfo";
 import { legalDocumentID } from "./documents/legalDocuments";
 import { compensationsId } from "./documents/compensations";
-import { companyLocationID } from "./documents/companyLocations";
+import { companyLocationID } from "./documents/companyLocation";
 
 export default (S: StructureBuilder) =>
   S.list()
